CLI/requirements.txt
elasticsearch
PyYAML
